feat(panel): Panel provider
Fixes #10006 - Allows users to be able to add the $mdPanelProvider to the config block
of their application. They can use this provider to create one or more
configurations for upcoming panel elements.

Fixes #10162 - Alters the configuration of a panel with a user-defined id that has been
saved when that panel is re-created. This will update the configuration properties such
as panelClass, etc.